requirements.txt
flask==3.0.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
openai==1.61.0
httpx==0.28.1
flask-limiter==3.5.1 